pyproject.toml
[tool.poetry]
name = "topos"
version = "0.2.1"
description = "The official Python client for Topos."
authors = ["Dialogues <info@dialogues.ai>"]
license = "MIT"
readme = "README.md"
homepage = "https://dialogues.ai"

[tool.poetry.dependencies]
python = "^3.9"
httpx = "^0.27.0"
fastapi = "0.109.2"
uvicorn = "0.20.0"
websockets = "11.0.3"
ollama = "0.1.9"
spacy = "3.7.2"
pydantic = "2.7.4"
transformers = "4.40.2"
torch = "2.3.0"
diffusers = "0.27.2"
accelerate = "0.30.1"
emoji = "0.6.0"
pyyaml = "6.0.1"
python-dotenv = "1.0.0"
openai = "1.30.4"
neo4j = "^5.20.0"
nltk = "^3.8.1"
scikit-learn = "^1.5.0"
sentence-transformers = "^3.0.1"
sentencepiece = "^0.2.0"
google = "^3.0.0"
protobuf = "^5.27.1"
matplotlib = "^3.9.0"
bertopic = "^0.16.2"
umap-learn = "^0.5.6"
pyjwt = "^2.8.0"
python-multipart = "^0.0.9"
pytest-asyncio = "^0.23.7"
textblob = "^0.18.0.post0"
pystray = "0.19.5"

supabase = "^2.6.0"
psycopg2-binary = "^2.9.9"
en-core-web-sm = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl"}
en-core-web-lg = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl"}
en-core-web-md = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl"}
en-core-web-trf = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.8.0/en_core_web_trf-3.8.0-py3-none-any.whl"}
aiokafka = "^0.11.0"
[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
pytest-cov = "^4.1.0"
pytest-httpserver = "^1.0.8"
pillow = "^10.2.0"
ruff = "^0.1.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
topos = "topos.cli:main"
    
